#49 refactor: 무너 3D 모델 파일 분리
